+++ assets/plugins/wwsu-audio/js/wwsu-silence.js
@@ -1,13 +1,14 @@
 /**
  * This class implements the silence detection system.
+ * It is expected to handle silence via the audioVolume event.
  *
  * @requires WWSUaudio The WWSUaudio class for making an input device
  * @requires WWSUevents WWSU event emitter.
  */
-class WWSUsilence extends WWSUevents {
+class WWSUsilenceaudio extends WWSUevents {
 	/**
 	 * Construct the class.
-	 * 
+	 *
 	 * @param {AudioContext} audioContext The audioContext to use (should use the one from wwsu-audio if possible)
 	 * @param {MediaStreamAudioDestinationNode} destination The destination node to use (should use the audioContext one)
 	 */
@@ -31,7 +32,9 @@ class WWSUsilence extends WWSUevents {
 
 				console.dir(this.destination.stream);
 
-				this.analyser = this.audioContext.createMediaStreamSource(this.destination.stream);
+				this.analyser = this.audioContext.createMediaStreamSource(
+					this.destination.stream
+				);
 
 				this.analyser
 					.connect(this.worklet)